+++ src/components/courses/CourseList.jsx
@@ -1,8 +1,9 @@
-import { Pagination, Skeleton } from 'antd'
+import { Pagination } from 'antd'
 import { useState } from 'react'
+import { useSelector } from 'react-redux'
 import styled from 'styled-components'
 
-import { CourseItem, CourseNotFound } from 'components/courses'
+import { CourseItem, NotFoundSearch } from 'components/courses'
 import { PageHeading, PageTitle } from 'components/shared'
 import { device } from 'styles/responsive'
 
@@ -27,6 +28,7 @@ const List = styled.ul`
 
 const CourseList = ({ courses }) => {
   const count = courses ? courses.length : 0
+  const { loading } = useSelector((state) => state.course)
 
   // pagination
   const [pageInfo, setPageInfo] = useState({
@@ -54,7 +56,7 @@ const CourseList = ({ courses }) => {
             <CourseItem data={data} key={data.id} />
           ))
         ) : (
-          <CourseNotFound />
+          <NotFoundSearch loading={loading} active />
         )}
       </List>
