--- src/components/courses/CourseNotFound.jsx
@@ -1,85 +1,0 @@
-import { Skeleton } from 'antd'
-import { useSelector } from 'react-redux'
-import { Link } from 'react-router-dom'
-import styled from 'styled-components'
-
-import NotFoundSearch from 'assets/images/NotFoundSearch.png'
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  color: #ffffff;
-`
-
-const Title = styled.h2`
-  max-width: 30rem;
-  margin: 1rem 0 0.5rem;
-  font-weight: 400;
-  font-size: 0.875rem;
-  text-align: center;
-  text-decoration: none;
-  letter-spacing: 1px;
-  color: ${({ theme }) => theme.darksecondary};
-`
-
-const Content = styled.p`
-  max-width: 22rem;
-  font-weight: 400;
-  font-size: 0.875rem;
-  text-align: center;
-  text-decoration: none;
-  letter-spacing: 1px;
-  color: ${({ theme }) => theme.secondary};
-`
-
-const StyledLink = styled(Link)`
-  font-weight: 400;
-  font-size: 0.75rem;
-  text-align: center;
-  text-decoration: none;
-  letter-spacing: 1px;
-  color: ${({ theme }) => theme.secondary};
-
-  &:hover {
-    text-decoration: underline;
-    text-decoration-thickness: 2px;
-    text-underline-offset: 1px;
-    color: ${({ theme }) => theme.secondary};
-  }
-`
-
-const Image = styled.img`
-  max-width: 12rem;
-  margin-top: 2rem;
-`
-
-const CourseNotFound = () => {
-  const { loading } = useSelector((state) => state.course)
-
-  return loading ? (
-    <Skeleton active />
-  ) : (
-    <Container>
-      <Image src={NotFoundSearch} alt="No results found" />
-      <Title>
-        <h1>No results found</h1>
-      </Title>
-
-      <Content>
-        Try adjusting your search or filter to find what you&rsquo;re looking
-        for.
-      </Content>
-
-      <Content style={{ fontSize: '0.75rem' }}>
-        Think this is an error? <StyledLink to="contact">Contact us</StyledLink>
-      </Content>
-    </Container>
-  )
-}
-
-export default CourseNotFound
