+++ src/pages/Courses.jsx
@@ -1,11 +1,9 @@
 import { Filter, X } from '@styled-icons/heroicons-outline'
 import { useEffect, useState } from 'react'
 import { Helmet } from 'react-helmet-async'
-import { useSelector } from 'react-redux'
 import styled from 'styled-components'
 
 import { CourseBody } from 'components/courses'
-import { LoaderAnimation } from 'components/shared'
 import { useViewportContext } from 'context/ViewportContext'
 import { breakpoints, device } from 'styles/responsive'
 
@@ -57,7 +55,6 @@ const Courses = () => {
         <title>Courses - ResoBin</title>
         <meta name="description" content="Courses availabe at IIT Bombay" />
       </Helmet>
-
       {Icon && (
         <IconContainer onClick={handleClick}>
           <Icon size="1.5rem" />
