+++ src/pages/Dashboard.jsx
@@ -5,9 +5,9 @@ import { Redirect, Route, Switch, useRouteMatch } from 'react-router-dom'
 import { Header } from 'components/header'
 import { MenuHorizontal, MenuVertical } from 'components/menu'
 import { LoaderAnimation } from 'components/shared'
+import Routes from 'config/Routes'
 import { useViewportContext } from 'context/ViewportContext'
 import { ScrollToTop } from 'hoc'
-import Routes from 'RoutesConfig'
 import { getCourseList } from 'store/courseSlice'
 import { breakpoints } from 'styles/responsive'
