+++ src/styles/scss/ant/_skeleton.scss
@@ -7,7 +7,7 @@
   &-content {
     padding: 1.5rem 1rem;
     border-radius: 8px;
-    background-color: $secondary-dark;
-    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.4);
+    background: rgba($secondary-dark, 0.4);
+    box-shadow: 0 0 0.5rem rgba(0, 0, 0, 0.2);
   }
 }
